legacy/requirements.txt
gensim==4.2.0
jmetalpy==1.5.5
language_tool_python==2.7.1
nltk==3.6.5
pygad==2.1.0
py-readability-metrics==1.4.5
Requests==2.28.1
